Grandes mudanças acontecerão

+++ MySqlNjs/app.js
@@ -12,7 +12,14 @@ app.use(express.static(__dirname + '/public'));
 app.use('/images', express.static(__dirname + '/public/images'));
 
 // Aqui diz básicamente que todos os arquivos que estão na pasta views são handlebars
-app.engine('handlebars', engine());
+app.engine('handlebars', engine({
+    helpers: {
+      // Função auxiliar para verificar igualdade (Valeu Ralf Lima)
+      conditionalEquality: function (parm1, parm2, options) {
+        return parm1 === parm2 ? options.fn(this) : options.inverse(this);
+      }
+    }
+  }));
 app.set('view engine', 'handlebars');
 app.set('views', './views');
 
@@ -35,6 +42,16 @@ connect.connect(function(error){
 
 // Rotas do site
 // ROTA PRINCIPAL
+app.get('/', function(req, res){
+    let sql = 'SELECT * FROM products';
+
+    connect.query(sql, function(error, ret){
+        if (error) throw error;
+        res.render('form', { products: ret });
+    });
+});
+
+// ROTA DE SITUAÇÃO, POR FAVOR NÃO CONFUNDIR COM A PRINCIPAL (Isso é para mim mesmo)
 app.get('/:situation', function(req, res){
     let sql = 'SELECT * FROM products';
 
@@ -46,41 +63,56 @@ app.get('/:situation', function(req, res){
 
 // Essa rota vai servir para cadastrar os produtos. OBS: coloquei algumas coisas em português, releve
 app.post('/register', function (req, res) {
-    let name = req.body.nome;
-    let value = req.body.valor;
-    let image = req.files.imagem.name;
+    try{
+        let name = req.body.nome;
+        let value = req.body.valor.trim();
+        let image = req.files.imagem.name;
 
-    let sql = `INSERT INTO products (nome, valor, imagem) VALUES ('${name}', ${value}, '${image}')`;
+        if(name == '' || value == '' || isNaN(value)){
+            res.redirect('/register-failed')
+        } else {
+            let sql = `INSERT INTO products (nome, valor, imagem) VALUES ('${name}', ${value}, '${image}')`;
 
-    connect.query(sql, function(error, ret){
-        if (error) throw error;
+            connect.query(sql, [name, parseFloat(value), image], function(error, ret){
+            if (error) throw error;
 
-        req.files.imagem.mv(__dirname + '/public/images/' + req.files.imagem.name, function(err) {
+            req.files.imagem.mv(__dirname + '/public/images/' + req.files.imagem.name, function(err) {
             if (err) throw err;
-            console.log('Image uploaded!');
+                console.log('Sucesso na imagem, meu brother');
         });
-        console.log(ret);
-    });
+            console.log(ret);
+        });
 
-    res.redirect('/');
+            res.redirect('/register-success');
+    } 
+
+    } catch (err) {
+        res.redirect('/register-failed');
+    }
 });
 
 // Essa rota vai remover os produtos e apaga-los do diretório
 app.get('/remove/:codigo&:imagem', function(req, res){
-    let sql = `DELETE FROM products WHERE codigo = ${req.params.codigo}`;
+    
+    try {
+        let sql = `DELETE FROM products WHERE codigo = ${req.params.codigo}`;
 
-    connect.query(sql, function(error, ret){
+        connect.query(sql, function(error, ret){
         if (error) throw error;
 
         fs.unlink(__dirname + '/public/images/' + req.params.imagem, (error_image) => {
             if (error_image) {
                 console.log("Deu erro em remover a imagem, meu chapa");
             } else {
-                console.log('Image deleted!');
+                console.log('Imagem deletada!');
             }
         });
-    });
-    res.redirect('/');
+      });
+        res.redirect('/remove-success');
+    } catch (error) {
+        res.redirect('/remove-failed');
+    }
+
 });
 
 app.get('/form-edit/:codigo', function(req, res){
@@ -100,43 +132,47 @@ app.post('/edit', function(req, res) {
     let cod = req.body.codigo;
     let nameImage = req.body.nameImage;
 
-    try {
-        if (req.files && req.files.imagem) {
-            let image = req.files.imagem;
+    if(name == '' || value == '' || isNaN(value)){
+        res.redirect('/edit-failed');
 
-            let sql = `UPDATE products SET nome = '${name}', valor = ${value}, imagem = '${image.name}' WHERE codigo = ${cod}`;
-
-            connect.query(sql, function(error, ret) {
-                if (error) throw error;
-
-                fs.unlink(__dirname + '/public/images/' + nameImage, (error_image) => {
-                    if (error_image) {
-                        console.log("Deu erro em remover a imagem, meu chapa");
-                    } else {
-                        console.log('Antiga imagem foi de ralo!');
-                    }
+    } else {
+        try {
+            if (req.files && req.files.imagem) {
+                let image = req.files.imagem;
+    
+                let sql = `UPDATE products SET nome = '${name}', valor = ${value}, imagem = '${image.name}' WHERE codigo = ${cod}`;
+    
+                connect.query(sql, function(error, ret) {
+                    if (error) throw error;
+    
+                    fs.unlink(__dirname + '/public/images/' + nameImage, (error_image) => {
+                        if (error_image) {
+                            console.log("Deu erro em remover a imagem, meu chapa");
+                        } else {
+                            console.log('Antiga imagem foi de ralo!');
+                        }
+                    });
+    
+                    image.mv(__dirname + '/public/images/' + image.name, function(err) {
+                        if (err) throw err;
+                        console.log('Novo meliante acionado!');
+                    });
+    
+                    console.log(ret);
                 });
-
-                image.mv(__dirname + '/public/images/' + image.name, function(err) {
-                    if (err) throw err;
-                    console.log('Novo meliante acionado!');
+            } else {
+                let sql = `UPDATE products SET nome = '${name}', valor = ${value} WHERE codigo = ${cod}`;
+    
+                connect.query(sql, function(error, ret) {
+                    if (error) throw error;
+                    console.log(ret);
                 });
-
-                console.log(ret);
-            });
-        } else {
-            let sql = `UPDATE products SET nome = '${name}', valor = ${value} WHERE codigo = ${cod}`;
-
-            connect.query(sql, function(error, ret) {
-                if (error) throw error;
-                console.log(ret);
-            });
+            }
+        } catch (error) {
+            console.error('Error:', error);
         }
-    } catch (error) {
-        console.error('Error:', error);
+        res.redirect('/edit-success');
     }
-
-    res.redirect('/');
 });
 
 app.listen(8080);
